Fix weather cold-miss and numpy overflow in RMS calculation
Weather: retry pre-warm 3x on startup, sync fetch on cold miss so LLM
always has weather data when user asks. Stale cache still served instantly
with async background refresh.

RMS: cast int16 audio to float64 upfront to prevent overflow when squaring
(int16 max 32767 squared exceeds int16 range, causing sqrt of negative).

+++ jf_sebastian/utils/audio_utils.py
@@ -37,8 +37,8 @@ def calculate_rms(audio_data: bytes, window_ms: int = 100, sample_rate: int = 16
         >>> if peak_rms < settings.MIN_AUDIO_RMS:
         ...     print("No speech detected in any window")
     """
-    # Convert bytes to numpy array of 16-bit integers
-    audio_array = np.frombuffer(audio_data, dtype=np.int16)
+    # Convert bytes to float64 array (avoids int16 overflow when squaring)
+    audio_array = np.frombuffer(audio_data, dtype=np.int16).astype(np.float64)
 
     # Calculate window size in samples
     window_samples = int(sample_rate * window_ms / 1000)
